refactor(backend): reorganize domain packages
