+++ opicer-api/src/main/java/com/opicer/api/auth/application/JwtService.java
@@ -1,8 +1,9 @@
-package com.opicer.api.auth;
+package com.opicer.api.auth.application;
 
+import com.opicer.api.auth.domain.AuthUserPrincipal;
 import com.opicer.api.config.AuthProperties;
-import com.opicer.api.user.AuthProvider;
-import com.opicer.api.user.UserRole;
+import com.opicer.api.user.domain.AuthProvider;
+import com.opicer.api.user.domain.UserRole;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
