+++ opicer-api/src/main/java/com/opicer/api/auth/domain/AuthUserPrincipal.java
@@ -1,7 +1,7 @@
-package com.opicer.api.auth;
+package com.opicer.api.auth.domain;
 
-import com.opicer.api.user.AuthProvider;
-import com.opicer.api.user.UserRole;
+import com.opicer.api.user.domain.AuthProvider;
+import com.opicer.api.user.domain.UserRole;
 import java.util.UUID;
 
 public record AuthUserPrincipal(
