+++ opicer-api/src/main/java/com/opicer/api/auth/infrastructure/OAuth2LoginFailureHandler.java
@@ -1,4 +1,4 @@
-package com.opicer.api.auth;
+package com.opicer.api.auth.infrastructure;
 
 import com.opicer.api.config.AuthProperties;
 import jakarta.servlet.http.HttpServletRequest;
