+++ opicer-api/src/main/java/com/opicer/api/user/application/UserService.java
@@ -1,5 +1,9 @@
-package com.opicer.api.user;
+package com.opicer.api.user.application;
 
+import com.opicer.api.user.domain.AuthProvider;
+import com.opicer.api.user.domain.User;
+import com.opicer.api.user.domain.UserRole;
+import com.opicer.api.user.infrastructure.UserRepository;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
