+++ opicer-api/src/main/java/com/opicer/api/user/domain/AuthProvider.java
@@ -1,4 +1,4 @@
-package com.opicer.api.user;
+package com.opicer.api.user.domain;
 
 public enum AuthProvider {
 	KAKAO
