+++ opicer-api/src/main/java/com/opicer/api/user/domain/User.java
@@ -1,4 +1,4 @@
-package com.opicer.api.user;
+package com.opicer.api.user.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
