+++ opicer-api/src/test/java/com/opicer/api/auth/application/JwtServiceTest.java
@@ -1,8 +1,9 @@
-package com.opicer.api.auth;
+package com.opicer.api.auth.application;
 
+import com.opicer.api.auth.domain.AuthUserPrincipal;
 import com.opicer.api.config.AuthProperties;
-import com.opicer.api.user.AuthProvider;
-import com.opicer.api.user.UserRole;
+import com.opicer.api.user.domain.AuthProvider;
+import com.opicer.api.user.domain.UserRole;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
